+++ Yanni/activationFunctionsCuda.cu
@@ -115,7 +115,7 @@ inline float SoftMaxDerivativeCuda(float& x, float* inputs, int* indexVector, in
 
 inline float SigmoidCuda(float& x)
 {
-	return  1.0 / (1.0 + exp(-x));
+	return  1.0f / (1.0f + expf(-x));
 }
 
 inline float SigmoidDerivativeCuda(float& x)
